onSaveStats and onLoadStats seem to be required in the relic class itself.
