+++ src/main/java/stsjorbsmod/relics/AlchemistsFireRelic.java
@@ -1,6 +1,7 @@
 package stsjorbsmod.relics;
 
 import com.evacipated.cardcrawl.mod.stslib.relics.ClickableRelic;
+import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
@@ -38,13 +39,6 @@ public class AlchemistsFireRelic extends CustomJorbsModRelic implements Clickabl
         this.flash();
     }
 
-    @Override
-    public AbstractRelic makeCopy() {
-        AlchemistsFireRelic copy = new AlchemistsFireRelic();
-        copy.stats = this.stats;
-        return copy;
-    }
-
     public int calculateBurningAmount(int amount) {
         return (amount * (100 - BURNING_FALLOFF_RATE)) / 100;
     }
@@ -80,15 +74,21 @@ public class AlchemistsFireRelic extends CustomJorbsModRelic implements Clickabl
     }
 
     @Override
-    public Object getStatsToSave() {
+    public JsonElement onSaveStats() {
+        // An array makes more sense if you want to store more than one stat
+        Gson gson = new Gson();
         ArrayList<Integer> statsToSave = new ArrayList<>();
         statsToSave.add(stats.get(STAT_BURNING_KEPT));
-        return statsToSave;
+        return gson.toJsonTree(statsToSave);
     }
 
     @Override
-    public void setStatsOnLoad(JsonElement jsonElement) {
-        JsonArray jsonArray = jsonElement.getAsJsonArray();
-        stats.put(STAT_BURNING_KEPT, jsonArray.get(0).getAsInt());
+    public void onLoadStats(JsonElement jsonElement) {
+        if (jsonElement != null) {
+            JsonArray jsonArray = jsonElement.getAsJsonArray();
+            stats.put(STAT_BURNING_KEPT, jsonArray.get(0).getAsInt());
+        } else {
+            resetStats();
+        }
     }
 }
