+++ src/main/java/stsjorbsmod/relics/CustomJorbsModRelic.java
@@ -3,11 +3,15 @@ package stsjorbsmod.relics;
 import basemod.abstracts.CustomRelic;
 import com.badlogic.gdx.graphics.Texture;
 import com.megacrit.cardcrawl.cards.AbstractCard.CardColor;
+import com.megacrit.cardcrawl.relics.AbstractRelic;
 import stsjorbsmod.JorbsMod;
 import stsjorbsmod.util.TextureLoader;
 
+import java.util.HashMap;
+
 public abstract class CustomJorbsModRelic extends CustomRelic {
     public final CardColor relicColor;
+    protected final HashMap<String, Integer> stats = new HashMap<>();
 
     private static Texture relicTextureFromId(String id) {
         String unprefixedId = id.replace(JorbsMod.MOD_ID + ":","");
@@ -37,5 +41,12 @@ public abstract class CustomJorbsModRelic extends CustomRelic {
     @Override
     public String getUpdatedDescription() {
         return DESCRIPTIONS[0].replaceAll(JorbsMod.MOD_ID + ":", "#y");
+    }
+
+    @Override
+    public AbstractRelic makeCopy() {
+        CustomJorbsModRelic copy = (CustomJorbsModRelic) super.makeCopy();
+        copy.stats.putAll(stats);
+        return copy;
     }
 }
